refactoring server actions and services in related folders
