+++ src/app/[locale]/(app)/groups/[groupId]/(admin)/deed-templates/[deedTemplateId]/_components/CreateDeedStatus.tsx
@@ -4,7 +4,7 @@ import { Drawer, DrawerContent, DrawerTrigger } from "@/components/ui/drawer";
 import { DeedStatusForm } from "@/features/deeds/components/DeedStatusForm";
 import { useDeedTemplateById } from "@/features/deeds/hooks/useDeedTemplateById";
 import type { DeedStatusInsert } from "@/features/deeds/models";
-import { createDeedStatus } from "@/features/deeds/service";
+import { createDeedStatus } from "@/features/deeds/actions/deeds";
 import { useMutation } from "@tanstack/react-query";
 import { Plus } from "lucide-react";
 import { useParams } from "next/navigation";
