+++ src/app/[locale]/(app)/groups/[groupId]/(admin)/deed-templates/_components/DraggableDeedTemplates.tsx
@@ -4,7 +4,7 @@ import { ListTile } from "@/components/ListTile";
 import { View } from "@/components/layout/View";
 import { useDeedTemplatesByGroupId } from "@/features/deeds/hooks/useDeedTemplatesByGroupId";
 import type { DeedTemplate } from "@/features/deeds/models";
-import { changeOrderDeedTemplates } from "@/features/deeds/service";
+import { changeOrderDeedTemplates } from "@/features/deeds/actions/deeds";
 import { routes } from "@/lib/routes";
 import { Link } from "@/navigation";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
