+++ src/app/[locale]/(app)/settings/_components/NotificationsPreference.tsx
@@ -5,7 +5,7 @@ import { Switch } from "@/components/ui/switch";
 import {
   registerPushNotifications,
   unregisterPushNotifications,
-} from "@/features/notifications/service";
+} from "@/features/notifications/services/notifications";
 import { Bell } from "lucide-react";
 import { useTranslations } from "next-intl";
 import { useState } from "react";
