+++ src/features/deeds/hooks/useMyDeedTemplates.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useQuery } from "@tanstack/react-query";
-import { getMyDeedTemplates } from "../service";
+import { getMyDeedTemplates } from "../actions/deeds";
 
 export const useMyDeedTemplates = () => {
   const query = useQuery({
