+++ src/features/deeds/hooks/useMyDeedsByDate.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useQuery } from "@tanstack/react-query";
-import { getMyDeedsByDate } from "../service";
+import { getMyDeedsByDate } from "../actions/deeds";
 
 export const useMyDeedsByDate = (date: Date) => {
   const query = useQuery({
