+++ src/features/groups/hooks/useGroupById.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useQuery } from "@tanstack/react-query";
-import { getGroupById } from "../service";
+import { getGroupById } from "../actions/groups";
 
 export const useGroupById = (groupId: string) => {
   const query = useQuery({
