+++ src/features/groups/hooks/useMyGroupInvitations.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useQuery } from "@tanstack/react-query";
-import { getMyGroupInvitations } from "../service";
+import { getMyGroupInvitations } from "../actions/groups";
 
 export const useMyGroupInvitations = () => {
   const query = useQuery({
